requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
asgiref==3.4.1
beautifulsoup4==4.10.0
Django==3.2
django-bootstrap3==15.0.0
django-bootstrap4==3.0.1
django-comments-xtd==2.9.3
django-contrib-comments==2.1.0
django-filter==21.1
django-js-asset==1.2.2
django-mptt==0.13.4
django-rest-multiple-models==2.1.3
django-threadedcomments==1.2
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
docutils==0.17.1
gunicorn==21.2.0
Pillow==8.3.2
PyJWT==2.3.0
pytz==2021.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.2

